requirements.txt
bcrypt==4.0.1
blinker==1.6.3
click==8.1.7
colorama==0.4.6
couchbase==4.1.8
Flask==2.3.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
Werkzeug==2.3.0
